+++ src/main/java/me/blvckbytes/bbtweaks/mechanic/common/TimeoutHandler.java
@@ -5,6 +5,6 @@ import me.blvckbytes.bbtweaks.mechanic.MechanicInstance;
 @FunctionalInterface
 public interface TimeoutHandler<InstanceType extends MechanicInstance> {
 
-  void handle(InstanceSession<InstanceType> session, int timeoutSeconds);
+  void handle(InstanceSession<InstanceType> session, long timeoutSeconds);
 
 }
