+++ src/main/java/me/blvckbytes/bbtweaks/mechanic/inv_move/InvMoveInstance.java
@@ -21,7 +21,7 @@ public class InvMoveInstance extends SISOInstance {
   }
 
   @Override
-  public boolean tick(int time) {
+  public boolean tick(long time) {
     if (time % 5 != 0)
       return true;
