+++ src/main/java/me/blvckbytes/bbtweaks/mechanic/sign_flipper/SignFlipperInstance.java
@@ -39,7 +39,7 @@ public class SignFlipperInstance extends SISOInstance {
   }
 
   @Override
-  public boolean tick(int time) {
+  public boolean tick(long time) {
     if (!isBlockLoaded(adjacentSignBlock))
       return true;
